--- src/imephu/salt/annotation/__init__.py
@@ -1,23 +1,0 @@
-from imephu.salt.annotation.observation import (
-    GeneralProperties,
-    MagnitudeRange,
-    Target,
-    hrs_observation_annotation,
-    rss_fabry_perot_observation_annotation,
-    rss_imaging_observation_annotation,
-    rss_longslit_observation_annotation,
-    rss_mos_observation_annotation,
-    salticam_observation_annotation,
-)
-
-__all__ = [
-    "GeneralProperties",
-    "MagnitudeRange",
-    "Target",
-    "hrs_observation_annotation",
-    "rss_fabry_perot_observation_annotation",
-    "rss_imaging_observation_annotation",
-    "rss_longslit_observation_annotation",
-    "rss_mos_observation_annotation",
-    "salticam_observation_annotation",
-]
