+++ tests/imephu/salt/test_salt_finder_chart/test_hrs_finder_chart.png
