+++ tests/imephu/salt/test_salt_finder_chart/test_rss_fabry_perot_finder_chart.png
