+++ tests/imephu/salt/test_salt_finder_chart/test_rss_imaging_observation_annotation_True_.png
