+++ tests/imephu/salt/test_salt_finder_chart/test_rss_longslit_finder_chart.png
