+++ tests/imephu/salt/test_salt_finder_chart/test_rss_mos_finder_chart.png
