+++ tests/imephu/salt/test_salt_finder_chart/test_salticam_finder_chart_True_.png
